--- frontend/src/Components/analysis/Analysis.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import './../../styling/Analysis.css'
-
-export default function Analysis() {
-    return (
-        <div className="query_analysis_panels">
-            Analysis Box
-        </div>
-    )
-}
