+++ frontend/src/Components/analysis/search_analysis/ArtistName.js
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default function ArtistName({artist, handleArtistSelect}) {
+    return (
+        <div 
+            className='artistName'
+            onClick={() => {handleArtistSelect(artist)}}
+        >
+            <p>{artist.name}</p>
+        </div>
+    )
+}
