+++ frontend/src/styling/Query.css
@@ -1,22 +1,32 @@
-.query_analysis_panels {
-    background-color: #DDD;
-    height: 20%;
-    width: 50%;
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    /* justify-content: center; */
-    margin: 20px;
-    border-radius: 10px;
-}
 
 #queryWindow {
     display: flex;
     flex-direction: column;
-    justify-content: space-around;
+    justify-content: flex-start;
+    align-items: center;
+    margin: 20px;
+    border-radius: 10px;
+    background-color: #DDD;
+    height: auto;
+    width: 50%;
 }
 
 #results {
     display: flex;
     flex-direction: column;
+}
+
+.artistName {
+    height: 30px;
+    width: 300px;
+    border: 1px solid black;
+    margin-top: -1px;
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+}
+
+.artistName:hover {
+    background-color: #137c38;
 }
